Attempt at normalization, read description
I'm attempting normalization for first form here, however as indicated in the document,  I want to go over with Mike what proper normalization would be and if I'm understanding the concept clearly. Once we are clear on normalization concepts, we should be on track with our plan for the database creation.